Fix Issue #1, turbine display is using the Reactor buffer size (10M RF) instead of the Turbine buffer size (1M RF).
I know I had fixed this, but I only fixed it for the overall status monitor (the one with the summary for every device) and not for the Turbine displays. Passively cooled Reactors may store up to 10,000,000 RF while Turbines may only store up to 1,000,000 RF. :) I was also lazy and used the same function, getDeviceStoredEnergyBufferPercent(device), for both Reactors and Turbines. I've now split it into getReactorStoredEnergyBufferPercent and getTurbineStoredEnergyBufferPercent and called accordingly.

Sorry for the delay in fixing.