GUI for monitor configuration
Click on the new < > in the second monitor line to switch within a
device class, click the ---------- to switch device class. I'm not
entirely sure how obvious to the user that latter is, and I'd hate for
anyone to find out only by accident. README time?

Also, started using log levels, as the debug spam was eating what I
wanted to see. That might see some extended use in the future.