Reworked event handling, including network changes
...which are now causing the system to reinitialize. That means you can
simply plug in a new monitor (reactors and turbine and use it, no reboot
necessary. Plug and Play, baby!

It started when I was hunting a bug where clicks simply didn't register
when configuring monitors, and I think I found the culprit in sleep().
Replaced it with wait() from the wiki, and because having two event poll
loops in the same process is a Bad Idea, I shuffled some code around.

Since the event handler now runs out of wait() called by main(), the
wrapping loops and parallel execution were unnecessary. That in turn
means we don't have to watch out for concurrency issues any more, and
THAT means we can happily update the UI from the event handler. The gain
in responsiveness is noticeable.